requirements.txt

flask==2.3.2
#multiprocessing==2.6.2.1
pdf2image==1.16.3
pytesseract==0.3.10
pandas==2.0.3
PyMuPDF==1.23.26
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
tabula-py==2.3.0
firebase-admin==5.0.3
joblib==1.2.0
tabula==1.0.5
google-api-core==2.11.1
google-cloud-vision==3.4.4
#fitz==0.0.1.dev2

ci-info==0.3.0
configobj==5.0.8 
configparser==6.0.0 
etelemetry==0.3.0 

future==0.18.3 
isodate==0.6.1 
looseversion==1.3.0
nibabel==5.1.0 
nipype==1.8.6  
prov==2.0.0 
pydot==1.4.2 
pyxnat==1.6
rdflib==7.0.0 
simplejson==3.19.1 
traits==6.3.2

hypercorn==0.14.4
python-dotenv==1.0.0

